Setup config for test env to simplify logging
Adds an environment based configuration setup to allow running a
different logger in test and development.

The test logger skips logging for anything in the `/css` path and uses a
simpler format to make it easier to follow the output of browser tests.
It also writes to the (T).Log function so failed tests only show
relevant logged output.